Update LZ-4-2.adoc
Changed this learning goal to clarify that architects are not responsible for defining and documenting requirements. They should clarify the requirements and can support those responsible for the requirements, but as soon as the architects document requirements in the form of scenarios, quality tree etc., we are in effect defining requirements. It is not our responsibility to define or document how fast, reliable, secure, etc. the system should be.
Refer to Issue #15