02-Design Terms: Removed "functional and technical architectures"
As per Issue #263 removed the terms "fachliche und technische Architektur" / "functional and technical architecture" from the terms list.

I think that it's very important to differentiate between these two types of architecture, but there needs to be something more specific in the curriculum to handle this topic. Even if we do, it should be an R3 topics.